made some changes

+++ Shared/DataTransferObjects/BeverageDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace Shared.DataTransferObjects
 {
-    public record BeverageDto(int Id,string Name, decimal Price, int Count, string ImageUrl);
+    public record BeverageDto(int Id,string Name, int Price, int Count, string ImageUrl);
 }
